Convert mwm to use loops to store windows
Also, tags now represent their focus with a pointer to the loop of their
windows.

In order to implement this cleanly, tags now store a list of their windows and
I have removed all parts relating to tag masks.

This means that for now, there will be no support for tags visible on multiple
windows or multiple tags visible at the same time. I have some ideas about how
this could be implemented, but for now, I'd rather go with simpler code, as I
do not use this feature.